Move controllers and deps to parent view for TodoDetails
